Refactoring link upper banner : put NTC usage button class in a separate file
